backend/requirements.txt
# Mediloon Backend - Ultra Minimal (Pinecone + OpenRouter)
# Total size: ~50MB (vs 4GB with ChromaDB/torch)

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
aiosqlite>=0.19.0

# Environment
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0

# Vector store (cloud - no local ML packages!)
pinecone>=5.0.0

# Observability
langfuse>=2.0.0
